Fix inEdgeCount bug.

+++ Weighted_Page_Rank/src/com/yullage/bigdata2013/PageRankVertex.java
@@ -126,17 +126,17 @@ public class PageRankVertex extends Vertex<Text, NullWritable, PageRankWritable>
 	}
 
 	private void calculateWeight(Iterable<PageRankWritable> messages) {
-		double totalInCount = 0;
-		double totalOutCount = 0;
+		long totalInCount = 0;
+		long totalOutCount = 0;
 
-		Map<Text, double[]> edgeCountMap = new HashMap<Text, double[]>();
+		Map<Text, long[]> edgeCountMap = new HashMap<Text, long[]>();
 		for (PageRankWritable msg : messages) {
 			System.out.println("Sender ID = " + msg.getSenderId());
 			totalInCount += msg.getInEdgeCount().get();
 			totalOutCount += msg.getOutEdgeCount().get();
 			System.out.println("total in = " + totalInCount + "; total out = " + totalOutCount);
 
-			double[] edgeCounts = new double[2];
+			long[] edgeCounts = new long[2];
 			edgeCounts[0] = msg.getInEdgeCount().get();
 			edgeCounts[1] = msg.getOutEdgeCount().get();
 			System.out.println("in = " + edgeCounts[0] + "; out = " + edgeCounts[1]);
@@ -144,11 +144,11 @@ public class PageRankVertex extends Vertex<Text, NullWritable, PageRankWritable>
 		}
 
 		MapWritable weightMap = new MapWritable();
-		for (Entry<Text, double[]> entry : edgeCountMap.entrySet()) {
+		for (Entry<Text, long[]> entry : edgeCountMap.entrySet()) {
 			Text key = entry.getKey();
-			double[] value = entry.getValue();
+			long[] value = entry.getValue();
 			System.out.println("value in = " + value[0] + "; value out = " + value[1]);
-			double weight = (value[0] / totalInCount) * (value[1] / totalOutCount);
+			double weight = (value[0] / (double) totalInCount) * (value[1] / (double) totalOutCount);
 			weightMap.put(key, new DoubleWritable(weight));
 			System.out.println("Sender = " + key + "; Weight = " + weight);
 		}
